refactor(runtime): extract lifecycle producers
