+++ dharma_swarm/orchestrator.py
@@ -35,7 +35,7 @@ from dharma_swarm.models import (
     TopologyType,
     _new_id,
 )
-from dharma_swarm.runtime_state import ArtifactRecord, DelegationRun, TaskClaim
+from dharma_swarm.runtime_lifecycle import RuntimeLifecycle
 from dharma_swarm.runtime_contract import RuntimeEnvelope, RuntimeEventType
 from dharma_swarm.session_ledger import SessionLedger
 from dharma_swarm.sheaf import (
@@ -112,6 +112,7 @@ class Orchestrator:
             session_id=session_id,
             runtime_db_path=resolved_runtime_db_path,
         )
+        self._runtime_lifecycle = RuntimeLifecycle(self._ledger)
         self._telic_seam = self._init_telic_seam()
         self._shared_dir = shared_dir or self._derive_runtime_artifact_dir("shared")
         self._stigmergy_dir = stigmergy_dir or self._derive_runtime_artifact_dir("stigmergy")
@@ -655,198 +656,6 @@ class Orchestrator:
         if task is None or not isinstance(task.metadata, dict):
             return {}
         return dict(task.metadata)
-
-    @staticmethod
-    def _utc_datetime_from(value: Any, fallback: datetime | None = None) -> datetime:
-        if isinstance(value, datetime):
-            return value if value.tzinfo is not None else value.replace(tzinfo=timezone.utc)
-        if isinstance(value, (int, float)):
-            try:
-                return datetime.fromtimestamp(float(value), timezone.utc)
-            except Exception:
-                pass
-        if isinstance(value, str) and value.strip():
-            try:
-                parsed = datetime.fromisoformat(value.replace("Z", "+00:00"))
-                return parsed if parsed.tzinfo is not None else parsed.replace(tzinfo=timezone.utc)
-            except Exception:
-                pass
-        return fallback or datetime.now(timezone.utc)
-
-    @staticmethod
-    def _runtime_status_time(status: str) -> datetime | None:
-        if status in {"running", "completed", "failed"}:
-            return datetime.now(timezone.utc)
-        return None
-
-    def _runtime_state_store(self) -> Any | None:
-        return getattr(self._ledger, "_runtime_state", None)
-
-    def _ensure_runtime_run_id(self, td: TaskDispatch) -> str:
-        existing = str(td.metadata.get("runtime_run_id", "") or "").strip()
-        if existing:
-            return existing
-        run_id = f"run_{_new_id()}"
-        td.metadata["runtime_run_id"] = run_id
-        return run_id
-
-    def _runtime_metadata(
-        self,
-        td: TaskDispatch,
-        *,
-        status: str,
-        failure_code: str = "",
-        error: str = "",
-        result: str | None = None,
-    ) -> dict[str, Any]:
-        metadata = {
-            "source": "orchestrator",
-            "status": status,
-            "topology": td.topology.value if td.topology else "dispatch",
-            "timeout_seconds": td.timeout_seconds,
-            "retry_count": td.metadata.get("retry_count", 0),
-            "max_retries": td.metadata.get("max_retries", 0),
-            "claim_timeout_seconds": td.metadata.get("claim_timeout_seconds", 0),
-        }
-        if failure_code:
-            metadata["failure_code"] = failure_code
-        if error:
-            metadata["error"] = error[:500]
-        if result is not None:
-            metadata["result_chars"] = len(result or "")
-        return metadata
-
-    async def _record_runtime_task_claim(
-        self,
-        td: TaskDispatch,
-        *,
-        task: Task | None,
-        status: str,
-        failure_code: str = "",
-        error: str = "",
-    ) -> None:
-        store = self._runtime_state_store()
-        claim_id = str(td.metadata.get("claim_id", "") or "").strip()
-        if store is None or not claim_id:
-            return
-        task_meta = self._task_meta(task)
-        active_claim = task_meta.get("active_claim")
-        if not isinstance(active_claim, dict):
-            active_claim = {}
-        now = datetime.now(timezone.utc)
-        acked_at = self._runtime_status_time(status)
-        heartbeat_at = now if status in {"running", "completed", "failed"} else None
-        stale_after = None
-        if active_claim.get("claim_expires_at_epoch") is not None:
-            stale_after = self._utc_datetime_from(active_claim.get("claim_expires_at_epoch"))
-        claim = TaskClaim(
-            claim_id=claim_id,
-            task_id=td.task_id,
-            agent_id=td.agent_id,
-            status=status,
-            session_id=self._ledger.session_id,
-            claimed_at=self._utc_datetime_from(active_claim.get("claimed_at"), now),
-            acked_at=acked_at,
-            heartbeat_at=heartbeat_at,
-            stale_after=stale_after,
-            retry_count=max(0, self._coerce_int(td.metadata.get("retry_count"), 0)),
-            metadata=self._runtime_metadata(
-                td,
-                status=status,
-                failure_code=failure_code,
-                error=error,
-            ),
-        )
-        try:
-            await store.record_task_claim(claim)
-        except Exception:
-            logger.debug("Runtime task claim recording failed", exc_info=True)
-
-    async def _record_runtime_delegation_run(
-        self,
-        td: TaskDispatch,
-        *,
-        task: Task | None,
-        status: str,
-        failure_code: str = "",
-        error: str = "",
-        result: str | None = None,
-    ) -> None:
-        store = self._runtime_state_store()
-        if store is None:
-            return
-        run_id = self._ensure_runtime_run_id(td)
-        started_raw = td.metadata.get("runtime_run_started_at")
-        started_at = self._utc_datetime_from(started_raw)
-        if not started_raw:
-            td.metadata["runtime_run_started_at"] = started_at.isoformat()
-        task_meta = self._task_meta(task)
-        requested_output = task_meta.get("requested_output", [])
-        if not isinstance(requested_output, list):
-            requested_output = []
-        completed_at = datetime.now(timezone.utc) if status in {"completed", "failed"} else None
-        run = DelegationRun(
-            run_id=run_id,
-            task_id=td.task_id,
-            assigned_to=td.agent_id,
-            status=status,
-            session_id=self._ledger.session_id,
-            claim_id=str(td.metadata.get("claim_id", "") or ""),
-            parent_run_id=str(task_meta.get("parent_run_id", "") or ""),
-            assigned_by="orchestrator",
-            requested_output=[str(item) for item in requested_output],
-            current_artifact_id=str(task_meta.get("current_artifact_id", "") or ""),
-            started_at=started_at,
-            completed_at=completed_at,
-            failure_code=failure_code,
-            metadata=self._runtime_metadata(
-                td,
-                status=status,
-                failure_code=failure_code,
-                error=error,
-                result=result,
-            ),
-        )
-        try:
-            await store.record_delegation_run(run)
-        except Exception:
-            logger.debug("Runtime delegation run recording failed", exc_info=True)
-
-    async def _record_runtime_artifact(
-        self,
-        *,
-        task: Task,
-        artifact_id: str,
-        artifact_kind: str,
-        payload_path: Path,
-        checksum: str,
-        manifest_path: Path | None = None,
-        run_id: str = "",
-        metadata: dict[str, Any] | None = None,
-    ) -> None:
-        store = self._runtime_state_store()
-        if store is None:
-            return
-        artifact = ArtifactRecord(
-            artifact_id=artifact_id,
-            artifact_kind=artifact_kind,
-            session_id=self._ledger.session_id,
-            task_id=task.id,
-            run_id=run_id,
-            manifest_path=str(manifest_path or ""),
-            payload_path=str(payload_path),
-            checksum=checksum,
-            promotion_state="ephemeral",
-            metadata={
-                "source": "orchestrator._persist_result",
-                "task_title": task.title,
-                **(metadata or {}),
-            },
-        )
-        try:
-            await store.record_artifact(artifact)
-        except Exception:
-            logger.debug("Runtime artifact recording failed", exc_info=True)
 
     def _resolve_timeout_seconds(self, task: Task | None, fallback: float) -> float:
         meta = self._task_meta(task)
@@ -1776,14 +1585,14 @@ class Orchestrator:
             max_retries=max_retries,
             backoff=backoff,
         )
-        await self._record_runtime_task_claim(
+        await self._runtime_lifecycle.record_task_claim(
             td,
             task=task,
             status="failed",
             failure_code=failure_class,
             error=error,
         )
-        await self._record_runtime_delegation_run(
+        await self._runtime_lifecycle.record_delegation_run(
             td,
             task=task,
             status="failed",
@@ -2039,7 +1848,7 @@ class Orchestrator:
         )
         logger.info("_assign_dispatch(%s): update_task=%.2fs", td.task_id[:8], _adt.monotonic() - _pe_t1)
         self._active_dispatches[td.task_id] = td
-        await self._record_runtime_task_claim(
+        await self._runtime_lifecycle.record_task_claim(
             td,
             task=task_for_gate,
             status="claimed",
@@ -2127,7 +1936,7 @@ class Orchestrator:
                 status=TaskStatus.RUNNING,
                 metadata=run_meta,
             )
-            await self._record_runtime_task_claim(
+            await self._runtime_lifecycle.record_task_claim(
                 td,
                 task=task,
                 status="running",
@@ -2182,7 +1991,7 @@ class Orchestrator:
             self._coerce_float(td.timeout_seconds, self._default_timeout_seconds),
         )
         try:
-            await self._record_runtime_delegation_run(
+            await self._runtime_lifecycle.record_delegation_run(
                 td,
                 task=task,
                 status="running",
@@ -2255,12 +2064,12 @@ class Orchestrator:
                 self._yoga.record_completion(td.agent_id)
             logger.info("Task %s completed by agent %s", td.task_id, td.agent_id)
             duration_sec = max(0.0, time.monotonic() - run_started)
-            await self._record_runtime_task_claim(
+            await self._runtime_lifecycle.record_task_claim(
                 td,
                 task=task,
                 status="completed",
             )
-            await self._record_runtime_delegation_run(
+            await self._runtime_lifecycle.record_delegation_run(
                 td,
                 task=task,
                 status="completed",
@@ -2602,7 +2411,7 @@ class Orchestrator:
             logger.debug("Shared artifact write failed (non-fatal): %s", exc)
 
         if shared_artifact is not None:
-            await self._record_runtime_artifact(
+            await self._runtime_lifecycle.record_artifact(
                 task=task,
                 artifact_id=f"artifact_task_result_{task.id}",
                 artifact_kind="task_result",
